+++ src/availability.ts
@@ -1,4 +1,4 @@
-import { store } from "./persistent";
+import { getStore } from "./persistent";
 
 let tagDisabled: string[] | undefined;
 let tagEnabled: string[] | undefined;
@@ -8,13 +8,13 @@ export function isLog(tag: string | undefined, silent?: boolean): boolean {
     if (!tag) return true;
     if (getTagDisabled(tag)) return false;
     if (getTagEnabled(tag)) return true;
-    if (store.config.value.t0.includes(tag)) return false;
-    if (store.config.value.t1.includes(tag)) return true;
+    if (getStore().config.value.t0.includes(tag)) return false;
+    if (getStore().config.value.t1.includes(tag)) return true;
     return !silent;
 }
 
 export function isVariable(): boolean {
-    return !(getVariableDisabled() || store.config.value.v0);
+    return !(getVariableDisabled() || getStore().config.value.v0);
 }
 
 function getStorageValue<T>(
@@ -49,4 +49,3 @@ function getTagEnabled(tag: string) {
     }
     return tagEnabled.includes(tag);
 }
-