+++ src/config.ts
@@ -1,13 +1,13 @@
-import { store } from "./persistent";
+import { getStore } from "./persistent";
 
 export function disableTags(tags: string[]) {
-    store.config.value.t0.push(...tags);
+    getStore().config.value.t0.push(...tags);
 }
 
 export function enableTags(tags: string[]) {
-    store.config.value.t1.push(...tags);
+    getStore().config.value.t1.push(...tags);
 }
 
 export function disableVariable() {
-    store.config.value.v0 = true;
+    getStore().config.value.v0 = true;
 }