+++ src/index.ts
@@ -1,9 +1,4 @@
 import { logMessage } from "./message";
 
-/**
- * @deprecated don't use this
- */
-const logConfig = { set: (..._: any[]) => {} };
-
 export * as LogConfig from "./config";
-export { logMessage as lm, logConfig, logMessage };
+export { logMessage as lm, logMessage };