+++ src/message.ts
@@ -1,10 +1,10 @@
 import { isLog, isVariable } from "./availability";
 import { getId } from "./id";
-import { store } from "./persistent";
+import { getStore } from "./persistent";
 import { getTag } from "./tag";
 
 type LogMessageReturn = { update: (...value: any[]) => void; remove: () => void };
-type LogMessageMeta = { tag?: string; bg?: string; fg?: string; silent?: boolean };
+type LogMessageMeta = { tag?: string; bg?: string; fg?: string; border?: string; silent?: boolean };
 
 export function logMessage(tag: string | undefined, ...value: any[]): LogMessageReturn;
 export function logMessage(meta: LogMessageMeta, ...value: any[]): LogMessageReturn;
@@ -16,6 +16,7 @@ export function logMessage(
     const tag = omode ? tagOrMeta.tag : tagOrMeta;
     const meta = omode ? tagOrMeta : undefined;
     const id = getId();
+    const store = getStore();
 
     if (isVariable()) store.messages.value.push({ id, tag, timestamp: Date.now(), value });
     if (isLog(tag, meta?.silent)) console.log(...getTag(tag, meta), ...value);